+++ src/commands/test.js
@@ -2,5 +2,5 @@ import karmaServer from '../karmaServer'
 
 export default function test(args, cb) {
   console.log('nwb: test')
-  karmaServer(args, {presets: ['react']}, cb)
+  karmaServer(args, cb)
 }